web/backend/requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
python-multipart>=0.0.12
python-dotenv>=1.2.1
pydantic>=2.11.0
openpyxl>=3.1.5
pandas>=2.3.3
pyyaml>=6.0.3
aiofiles>=24.1.0
